集成Shiro 优化页面美观

+++ SpringBoot-RateLimiter/src/main/java/com/dashuai/learning/ratelimiter/service/RateLimiterService.java
@@ -16,7 +16,7 @@ public class RateLimiterService {
     private RateLimiter rateLimiter;
 
 
-    public RateLimiterService(int count) {
+    public RateLimiterService(double count) {
         rateLimiter = RateLimiter.create(count);
     }
